dir structure, project info in the readme
