--- src/App.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-
-      </header>
-    </div>
-  );
-}
-
-export default App;
