+++ src/App/App.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import App from './App';
+
+test('renders App and all children', () => {
+  render(<App />);
+  const appElement = window.document.getElementsByClassName('App')[0];
+  expect(appElement).not.toBeUndefined();
+});
