+++ src/App/App.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Header from '../components/Header/Header';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Header />
+      <p>
+        Edit <code>src/App.tsx</code> and save to reload.
+      </p>
+    </div>
+  );
+}
+
+export default App;
