+++ src/components/Header/Header.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Header.css';
+
+function Header() {
+  return (
+    <header className="App-header">
+      <p>
+        This is the Family News Feed informational website!
+      </p>
+    </header>
+  );
+}
+
+export default Header;
